fix: rename builderView folser to builder
