+++ src/modules/builder/ImageDisplay.tsx
